03_ax_curriculum_chatbot/requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
rich>=13.7.0
